Use lsblk to attempt to guess partition's fs type
If not found it will fall back to other tools.

Note that fs_type function is not expected to return a non-zero error
code even if it failed to detect the type.

If filesystem is undetermined then "NOT-DETECTED" will be echoed.

This change will prevent mount hanging on extended partitions (#768902).

Signed-off-by: Philippe Coval <rzr@gna.org>
Tweaked-by: Colin Watson <cjwatson@debian.org>
Bug-Debian: 768902